Add pulseaudio to cerbero build system.
Pulseaudio v13 can now be built for windows including libsndfile 1.0.28
dependency. Mingw build is perfectly fine. MSVC build works fine
standalone. gst-plugin with MSVC is not working due to different CRT of
pulseaudio build. Currently plugins compiled against libraries, built
with autotools do not work in msvc build. To fix this, the cerbero meson build
has to be prepared for windows as well. This is a more complex endeavor,
because linking against mingw32 is not possible here to ensure full
windows compatibility.